requirements.txt
flask
# YOLOv7 requirements
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.11.0
torchvision>=0.12.0
tqdm>=4.64.0
seaborn>=0.11.0
pandas
firebase_admin
bs4
#deploy
gunicorn